Move everyting disk operations related to disk.py
